Extract the suspendable-child park handle into a shared helper
Driving a suspendable body -- spawn it, surface its approval and author
awaitSignal parks over a FIFO the caller drains via next(), relay grants
back, complete it -- is independent of how the body's env was built.
Separate it from createSidecarSpawnSuspendableChild's onTrigger env
assembly as createSuspendableChildHandle, so the handle machinery is the
single implementation behind the SuspendableChildHandle contract
whatever env a caller hands it.

The channel teardown (signalChannel.stop()) is not on the SignalChannel
interface, so it rides a caller-supplied cleanup hook rather than the
handle reaching for it on env.signalChannel.
