+++ apps/sidecar/src/workflow-substrate-factory.ts
@@ -27,7 +27,6 @@ import path from "node:path";
 
 import { type } from "arktype";
 
-import { signalName } from "@intx/types";
 import { InferenceSource } from "@intx/types/runtime";
 import type {
   ApprovalSnapshot,
@@ -104,6 +103,7 @@ import {
   baseStepId,
   collectDeclaredPluginNames,
   createNoopDrainController,
+  createSuspendableChildHandle,
   emptyState,
   rewriteInlineChildWorkflowBodies,
   runtimeRun,
@@ -112,7 +112,6 @@ import {
   type Scheduler,
   type StepInvokeRequest,
   type StepInvokeResult,
-  type SuspendableChildPark,
   type WorkflowAuthorizeFn,
   type WorkflowDefinition,
   type WorkflowRuntimeEnv,
@@ -1551,144 +1550,14 @@ export function createSidecarSpawnSuspendableChild(
         : {}),
     });
 
-    // FIFO the caller drains via `next()`: each entry is either an approval
-    // park to proxy up or a fatal illegal-park error. A single waiter slot
-    // suffices because `next()` has exactly one consumer (`runOnTrigger`)
-    // driving it sequentially, mirroring the signal channel's single-consumer
-    // shape.
-    type BodyEvent =
-      | { kind: "park"; park: SuspendableChildPark }
-      | { kind: "signal-park"; name: string }
-      | { kind: "error"; error: Error };
-    const events: BodyEvent[] = [];
-    let wake: (() => void) | null = null;
-    const notify = (): void => {
-      if (wake !== null) {
-        const resolve = wake;
-        wake = null;
-        resolve();
-      }
-    };
-
-    const env: WorkflowRuntimeEnv = {
-      ...baseEnv,
-      onPark: (park) => {
-        if (park.parkKind === "approval") {
-          events.push({
-            kind: "park",
-            park: {
-              correlationId: park.correlationId,
-              ...(park.approvalSnapshot !== undefined
-                ? { approvalSnapshot: park.approvalSnapshot }
-                : {}),
-            },
-          });
-        } else {
-          events.push({
-            kind: "error",
-            error: new Error(
-              `onTrigger body ${childRunId} parked on a control-plane input ` +
-                `channel (${park.correlationId}); a suspendable body may not ` +
-                `re-arm an input park -- the section proxies approvals only, ` +
-                `so this park has no resolver`,
-            ),
-          });
-        }
-        notify();
-      },
-      // A body `awaitSignal` gate on an author name: surface it so the section
-      // proxies it up as a signal-relay await and relays the resolved signal
-      // back via `deliverSignal`. Without this the body would park on the
-      // signal channel with nothing upstream to route a delivery to it.
-      onSignalPark: (park) => {
-        events.push({ kind: "signal-park", name: park.name });
-        notify();
-      },
-    };
-
-    // On resume, drive the run from its durable log; the body step re-parks
-    // silently (a re-park does not re-fire onPark), and the caller relays the
-    // grant via resume on the correlation it recovered from its own log. On a
-    // fresh spawn, seed the run with the event's trigger payload.
-    const handle = runtimeRun(
-      rewrittenDefinition,
-      env,
-      resumeFromEvents !== undefined
-        ? { runId: childRunId, resumeFromEvents }
-        : { runId: childRunId, triggerPayload: input },
-    );
-
-    // The resulting `CancelRequested` is written under the supervisor principal
-    // wired into this run's repo store (see `controlPlanePrincipal` above): the
-    // kind handler requires a supervisor signer for any cancel origin, so a
-    // workflow-process-signed cancel would be refused and a parent abort would
-    // surface as a failed rather than cancelled child.
-    const cancelOnAbort = (): void => {
-      void handle.cancel("supervisor-operator", "parent cancelled");
-    };
-    if (signal.aborted) {
-      cancelOnAbort();
-    } else {
-      signal.addEventListener("abort", cancelOnAbort, { once: true });
-    }
-
-    let settled: {
-      terminalStatus: "completed" | "failed" | "cancelled";
-    } | null = null;
-    let failure: Error | null = null;
-    void handle.complete
-      .then((result) => {
-        settled = { terminalStatus: result.terminalStatus };
-      })
-      .catch((cause) => {
-        failure = cause instanceof Error ? cause : new Error(String(cause));
-      })
-      .finally(() => {
-        signal.removeEventListener("abort", cancelOnAbort);
-        void signalChannel.stop();
-        notify();
-      });
-
-    return {
-      next: async () => {
-        for (;;) {
-          const event = events.shift();
-          if (event !== undefined) {
-            if (event.kind === "error") {
-              // The body re-armed an input park nothing will resolve. Cancel
-              // the child so its terminal (and the channel teardown tied to
-              // it) fires, then surface the error: the throw lands the section
-              // run's terminal via `runOnTrigger`'s `runPrimitiveSafe`.
-              void handle.cancel(
-                "supervisor-operator",
-                "onTrigger body re-armed an unsupported input park",
-              );
-              throw event.error;
-            }
-            if (event.kind === "signal-park") {
-              return { kind: "signal-park", name: event.name };
-            }
-            return { kind: "park", park: event.park };
-          }
-          if (failure !== null) throw failure;
-          if (settled !== null) {
-            return {
-              kind: "terminal",
-              terminalStatus: settled.terminalStatus,
-            };
-          }
-          await new Promise<void>((resolve) => {
-            wake = resolve;
-          });
-        }
-      },
-      resume: async (correlationId, decision) => {
-        await signalChannel.deliver(signalName(correlationId), decision);
-      },
-      deliverSignal: async (name, payload, signalId) => {
-        await signalChannel.deliver(name, payload, signalId);
-      },
-    };
+    return createSuspendableChildHandle(baseEnv, {
+      definition: rewrittenDefinition,
+      childRunId,
+      input,
+      ...(resumeFromEvents !== undefined ? { resumeFromEvents } : {}),
+      signal,
+      cleanup: () => signalChannel.stop(),
+    });
   };
 }
